+++ weblate/screenshots/forms.py
@@ -50,9 +50,9 @@ class ScreenshotForm(forms.ModelForm):
             "translation": LanguageChoiceField,
         }
 
-    def __init__(self, component, data=None, files=None, instance=None):
+    def __init__(self, component, data=None, files=None, instance=None, initial=None):
         self.component = component
-        super().__init__(data=data, files=files, instance=instance)
+        super().__init__(data=data, files=files, instance=instance, initial=initial)
         self.fields[
             "translation"
         ].queryset = component.translation_set.prefetch_related("language")